pyproject.toml
[project]
name = "augmented-fiction"
version = "0.1.0"
description = "AI-augmented writing assistant"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "typer>=0.13",
    "jinja2>=3.1",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0",
    "ebooklib>=0.18",
    "beautifulsoup4>=4.12",
    "nltk>=3.9",
    "lxml>=5.0",
    "openai>=1.0",
    "scikit-learn>=1.5",
    "mangum>=0.21.0",
]

[project.scripts]
af = "augmented_fiction.cli.main:app"
af-web = "augmented_fiction.web.app:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/augmented_fiction"]
